+++ api.planx.uk/modules/webhooks/service/lowcalSessionEvents/schema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { ValidatedRequestHandler } from "../../../shared/middleware/validate";
-import { ScheduledEventResponse } from "../../../hasura/metadata";
+import { ValidatedRequestHandler } from "../../../../shared/middleware/validate";
+import { ScheduledEventResponse } from "../../../../hasura/metadata";
 
 export const createSessionEventSchema = z.object({
   body: z.object({
