+++ api.planx.uk/modules/webhooks/service/sendNotification/types.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ValidatedRequestHandler } from "../../../shared/middleware/validate";
+import { ValidatedRequestHandler } from "../../../../shared/middleware/validate";
 import {
   bopsSubmissionSchema,
   emailSubmissionSchema,
